fastapi/requirements.txt
fastapi
uvicorn
prometheus-client
prometheus-fastapi-instrumentator
python-logging-loki